+++ tools/tags/actions/admintag.php
@@ -1,4 +1,5 @@
 <?php
+use YesWiki\Security\Controller\SecurityController;
 
 if (!defined("WIKINI_VERSION")) {
     die("acc&egrave;s direct interdit");
@@ -6,6 +7,9 @@ if (!defined("WIKINI_VERSION")) {
 
 if ($this->UserIsAdmin()) {
     if (isset($_GET['delete_tag'])) {
+        if ($this->services->get(SecurityController::class)->isWikiHibernated()) {
+            throw new \Exception(_t('WIKI_IN_HIBERNATION'));
+        }
         $sql = 'DELETE FROM '.$this->config['table_prefix'].'triples WHERE property="http://outils-reseaux.org/_vocabulary/tag" and id IN ('.mysqli_real_escape_string($this->dblink, $_GET['delete_tag']).')';
         $this->Query($sql);
     }
